# docs/openapi/admin.yaml
openapi: 3.0.3
info:
  title: DeliveryHub API
  version: v1
  description: See docs/api.md for the conventions (errors, pagination, authentication).
paths:
...
  /api/v1/orders:
    get:
      operationId: orders_list
      description: A paginated list for staff; subclasses filter in get_queryset().
      summary: Orders, newest first
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
      - in: query
        name: search
        schema:
          type: string
        description: Order number, customer name or phone
      - in: query
        name: source
        schema:
          type: string
      - in: query
        name: status
        schema:
          type: string
      tags:
      - orders
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOrderSummaryList'
          description: ''
...
components:
  schemas:
...
    OrderSummary:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        source:
          type: string
        created_at:
          type: string
          format: date-time
        customer_name:
          type: string
        phone:
          type: string
        total:
          type: integer
        items_count:
          type: integer
        client_id:
          type: integer
          nullable: true
      required:
      - client_id
      - created_at
      - customer_name
      - id
      - items_count
      - phone
      - source
      - status
      - total
...
    PaginatedOrderSummaryList:
      type: object
      required:
      - count
      - page
      - pages
      - results
      properties:
        count:
          type: integer
          example: 135
        page:
          type: integer
          example: 1
        pages:
          type: integer
          example: 7
        results:
          type: array
          items:
            $ref: '#/components/schemas/OrderSummary'
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid